==> R/data_wrangling/2_gaze_xsit_wrangle.Rmd <==
---
title: "Gaze-xsit Timecourse Wrangling"
author: "Kyle MacDonald"
output: html_document
---

```{r chunk opts, echo = F}
rm(list = ls())
knitr::opts_chunk$set(echo=T, warning=F, cache=F, message=F, sanitize = T)
```

This script munges and analyzes the data for a project exploring the association between allocation of attention and cross-situational word learning in more or less ambiguous contexts.

* Input: processed data spreadsheet, subject info spreadsheet, trial info
* Output: d (mastersheet with times, trial, subject, condition, and gazes)

```{r setup} 
library(magrittr)
library(stringr)
library(tidyverse)
theme_set(ggthemes::theme_few())

# path information
read_path <- "../../data/2_processed_data/"
write_path <- "../../data/3_tidy_data/"
```

Read data.

```{r read data}
d <- read_csv(paste0(read_path, "processed_adult_data.csv"))
trialinfo <- read_csv("../../data/gaze_xsit_trialinfo.csv")
subinfo <- read_csv("../../data/subinfo_adult.csv")
```

Add participant information. 

```{r}
# hacky way to deal with shift in subject naming convention after
# data collection had begun
tmp_subinfo <- subinfo %>% 
  filter(!is.na(old_subid)) %>% 
  rename(subid_good = subid,
         subid = old_subid) %>% 
  select(subid, subid_good)

d %<>% 
  mutate(subid = ifelse(subid == "0216166_03", "021616_03", subid)) %>% 
  left_join(tmp_subinfo, by = "subid") %>% 
  mutate(subid = ifelse(is.na(subid_good), subid, subid_good)) %>% 
  dplyr::select(-subid_good)

d %<>% left_join(subinfo, by = "subid")
```

Add trial information.

```{r}
d %<>% left_join(trialinfo, by = "stimulus")
```

Filter out unusable participants. We do this by calculating the percentage of non-looks: looking away from the screen entirely. Note that offscreen == TRUE if x is 0 and y is 1050, so higher values of `bad_looks` means more away looking. 

```{r}
d_screentime <- d %>%
  mutate(offscreen = case_when(
     (x == 0 & y == 1050) ~ TRUE, 
     is.na(x) ~ TRUE, 
     is.na(y) ~ TRUE,
     TRUE ~ FALSE)
  ) %>% 
  select(subid, t, offscreen) %>%
  group_by(subid) %>%
  summarise(bad_looks = sum(offscreen, na.rm = TRUE), 
            total_timeslices = n())

d_screentime %<>% mutate(percent_bad_looks = bad_looks / total_timeslices)
```

Filter out participants more than 2 standard dev away from mean screentime. 

```{r}
sd_ss <- sd(d_screentime$percent_bad_looks)
mean_ss <- mean(d_screentime$percent_bad_looks)

d %<>% left_join(d_screentime, by = "subid")

d$keep_et <- ifelse(d$percent_bad_looks > (2 * sd_ss + mean_ss), 
                 'exclude', 'include')
```

Make table of exclusions based on the away filter. 

```{r}
d %>% 
  select(subid, keep_et) %>% 
  distinct() %>% 
  count(keep_et)
```

Now we filter

```{r}
d %<>% filter(keep_et == "include") 
```

## More data wrangling

Here we label Areas of Interest (ss_looking_char) to be Left Picture, Right Picture, and the Center Face. We also add trial number variable.

```{r}
d %<>% mutate(condition = ifelse(condition == "nogaze", 
                                 "no-gaze", 
                                 condition))
```

Make "sanity check" plot of looks based on x/y coordinates.

```{r}
d.rois <- sample_n(d, size = 0.1 * nrow(d)) # get a sample so we don't plot all 200k plus rows

ggplot(aes(x = x, y = y), data = d.rois) +
  stat_density_2d(aes(fill = ..level..), geom = "polygon") +
  lims(x = c(0, 1980))
```

Bin each timeslice as looking to one of the ROIs or "away"

```{r ss_looking_char setup}
d %<>% 
  mutate(ss_looking_factor = case_when(
    x >= 100 & x <= 750 & y <= 700 & y >= 100 ~ "left",
    x >= 1170 & x <= 1710 & y <= 700 & y >= 100 ~ "right",
    x <= 1500 & x >= 500 & y >= 700 ~ "face",
    x == 0 & y == 1050 ~ "missing",
    TRUE ~ "away"
    ),
  ss_looking_char = as.character(ss_looking_factor)
  )
```

Check how well our AOI coding did. 

```{r}
d %>% 
  filter(subid == subid[3]) %>% 
  ggplot(aes(x = x, y = y)) + 
  geom_line(col = "darkgrey") + 
  geom_point(aes(col = ss_looking_char), alpha = .3, size = 3) + 
  facet_wrap(~shortname*trial.type, ncol = 8) + 
  xlim(c(0,1920)) + 
  ylim(c(0,1080)) +
  theme(legend.position = "top") +
  theme_bw()
```

Add trial numbers.

```{r}
trial_num_vect <- c(c("fam", "fam", "fam", "fam"), seq(1,16))

df_trial_nums <- d %>% 
  select(subid, shortname, t) %>% 
  filter(!is.na(shortname)) %>% # this removes one NA row from participant (051216_01)
  arrange(subid, t) %>% 
  select(-t) %>% 
  unique() %>% 
  group_by(subid) %>% # this allows you to add the trial vector for each participant
  mutate(trial_num = trial_num_vect)

d %<>% left_join(df_trial_nums, by = c("subid", "shortname"))
```

Add correct column for exposure and test trials. A time slice is correct if the participant looked at the object that was kept from exposure to test.

We also add a column encoding whether the participant looked at the gaze target during exposure trials. This is only relevant for the gaze condition. In the no-gaze condition, each time slice will be False.

```{r}
d %<>% mutate(correct = ifelse(target_object == ss_looking_char, TRUE, FALSE),
              correct_char = ifelse(target_object == ss_looking_char, "target", 
                                    ifelse(ss_looking_char == "face", "face",
                                           "distracter")),
              gaze_follow = ifelse(look == ss_looking_char, TRUE, FALSE))
```

## Diagnotstic plots

Next, let's make another sanity check plot of raw looking behavior over time

```{r}
d_plot <- d %>%
  filter(trial.type != "familiar", 
         ss_looking_char != 'away', 
         ss_looking_char != "missing",
         !is.na(x),
         !is.na(y)) %>% 
  mutate(x = ifelse(x < 1 | x > 1919, NA, x), 
         y = ifelse(y < 1 | y > 1079, NA, y))
  

ggplot(filter(d_plot, subid == "020816_01"), 
       aes(x = t.stim, y = x)) + 
  geom_line(col = "darkgrey") + 
  geom_point(aes(col = ss_looking_char), alpha = .3) + 
  facet_wrap(~trial_num*trial.type, ncol = 8) + 
  ylim(c(0,1920)) + 
  geom_hline(yintercept = 1920/2, lty = 2) +
  theme_bw() +
  theme(legend.position = "top")
```

## Interpolation

In this section of code, we "fill in" values that are missing because of tracker loss, but that we are reasonably confident where the participant is looking. 

We define the interpolation algorithm as...

```{r}

```

## Save tidy timecourse data

```{r}
write_csv(x = d, path = paste0(write_path, "gaze_xsit_tidy_timecourse.csv"))
```
