==> R/data_wrangling/3_gaze_xsit_summarise.Rmd <==
---
title: "Gaze-xsit Summarise"
author: "Kyle MacDonald"
output: html_document
---

```{r chunk opts, echo = F}
rm(list = ls())
knitr::opts_chunk$set(echo=T, warning=F, cache=F, message=F, sanitize = T)
```

This script creates a summary table of participant looking behavior in the Gaze-Xsit experiment (a project exploring the association between allocation of attention and cross-situational word learning in more or less ambiguous contexts).

```{r setup} 
source("../helper_functions/helper_functions.R")
read_path <- "../../data/3_tidy_data/"
write_path <- "../../data/3_tidy_data/"
```

Read data. 

```{r}
d <- read_csv(file = paste0(read_path, "gaze_xsit_tidy_timecourse.csv"))
```

Create trial name that encodes both number and trial type. 

```{r}
d %<>% mutate(trial_num_full = paste(trial_num, trial.type, sep = "_")) 
```

Remove familiar trials

```{r}
d %<>% filter(trial_num != "fam", trial.type != "familiar")
```

## Initial data checks

Get the number of participants in each condition.

```{r}
d %>% 
  select(subid, condition) %>% 
  unique() %>% 
  group_by(condition) %>% 
  summarise(n_subs = n())
```

Get the total number of trials

```{r}
d %>% 
  filter(trial.type != "familiar") %>% 
  select(subid, condition, trial.type, trial_num) %>% 
  unique() %>% 
  group_by(subid, condition, trial.type) %>% 
  count() %>% 
  pull(n) == 16
```

Extract subid, condition, and trial_number. This is useful for maintaing the integrity of this information during the data joins later in the script. 

```{r}
ss_info <- d %>% 
  select(subid, condition) %>%
  unique()
```

## Get proportion looking scores

Here we create a summary data frame for analysis and visualizations.

First, we need to get a crosstabs of counts for where participants were looking (face, target, distracter) during each trial. The target_object variable encodes the object that will appear again during test. It does not have to be the target of gaze. 

Note the use of the complete() function from the tidyr package for completing missing combinations of data. It turns implicitly missing values into explicitly missing values.

```{r}
# get count of looking to each picture
ss <- d %>% 
  filter(ss_looking_char %in% c("left", "right", "face")) %>%
  group_by(subid, trial_num_full, condition, ss_looking_char, look, correct) %>% 
  summarise(count_looks = n()) 

# complete the cases 
# this takes care of when there was no looking to one of the images
ss %<>% 
  ungroup() %>% 
  complete(subid, trial_num_full, ss_looking_char, 
           fill = list(count_looks = 0)) %>%
  arrange(subid) %>% 
  select(-condition) # here we remove condition since it has NAs after the complete function

# get the total number of time slices for each trial and
ss %<>% 
  group_by(subid, trial_num_full) %>% 
  mutate(total_looks = sum(count_looks)) %>% 
  filter(total_looks > 0) # remove trials without any looking data

# get proportion looking to face and objects during each trial
ss %<>% mutate(prop_looking = round(count_looks / total_looks, 2))
```

Get proportion correct and incorrect looking on exposure and test trials. Note that we only use looks to the objects here to score the proportion correct and gaze following. That is, we don't penalize participants for looks to the face AOI. Put another way, these are measures of how participants choose to distribute attention when they disengage from the face. 

```{r}
ss_correct <- d %>% 
  filter(ss_looking_char %in% c("left", "right")) %>%
  group_by(subid, trial_num_full) %>% 
  summarise(m_correct = mean(correct, na.rm = T) %>% round(digits = 2),
            m_gaze_follow = mean(gaze_follow, na.rm = T) %>% round(digits = 2))
```

Join the prop correct and gaze following with the prop looking data.

```{r}
ss %<>% left_join(., ss_correct, by = c("subid", "trial_num_full"))
```

Get the frequency of gaze shifts within a trial.

```{r}
num_shifts_df <- d %>% 
  filter(ss_looking_char %in% c("left", "right", "face"),
         trial_num != "fam", trial.type != "familiar", 
         t.stim <= 4) %>% 
  group_by(subid, trial_num_full) %>% 
  summarise(num_shifts = get_freq_gaze_shifts(ss_looking_char, 
                                              min_fixation_len = 150,
                                              ms_per_frame = 33))
```

Make histogram of number of shifts within each trial.

```{r}
qplot(num_shifts_df$num_shifts, bins = 10)
```

Join shifting information with the rest of the data.  

```{r}
ss %<>% left_join(., num_shifts_df, by = c("subid", "trial_num_full"))
```

## Get RTs for first shifts on exposure and test trials

Map the score function over each participants' data frame. This returns an RT and a shift-type for all first shifts after noun onset. (Note that this will take a couple of minutes to run.) 

```{r}
trial_scores <- d %>% 
  filter(ss_looking_char %in% c("left", "right", "face"),
         trial_num != "fam", trial.type != "familiar") %>% 
  split(.$subid) %>% 
  purrr::map_df(score_smi_participant)
```

Note that the noun and the gaze both started at the beginning of each trial, which makes the first shift accuracy and RT a little tricker to understand. I think we can still analyze the frequency of shifting or the proportion of time spent looking at the face vs. objects as an index of the effect of social cues on gaze patterns.

Join with the rest of the Trial-level data. 

```{r}
ss %<>% left_join(trial_scores, by = c("subid", "trial_num_full"))
```

Some final variable clean-up -- separate information for trial number and trial type, so we are not encoding two types of information in a single variable. 

```{r}
ss %<>% separate(col = trial_num_full, c("trial_num", "trial_type"), remove = F)
```

Join condtion information 

```{r}
ss %<>% left_join(., ss_info, by = "subid")
```

## Final data checks

Number of participipants in each conditions should be 15.

```{r}
ss %>% 
  ungroup() %>% 
  select(subid, condition) %>% 
  unique() %>% 
  group_by(condition) %>% 
  summarise(n_subs = n())
```

Total number of trials for each participant should be 16. 

```{r}
d %>% 
  filter(trial.type != "familiar") %>% 
  select(subid, condition, trial.type, trial_num) %>% 
  unique() %>% 
  group_by(subid, condition, trial.type) %>% 
  count() %>% 
  pull(n) == 16
```

Checks look good! 

## Write to file

```{r}
write_csv(ss, paste0(write_path, "gaze_xsit_tidy_trial_level.csv"))
```
